refactor: rename resolve_reference.py to resolve_address.py
