+++ src/common/utils/get_resolved_document_by_id.py
@@ -3,7 +3,7 @@ from typing import Callable
 from common.address import Address
 from common.exceptions import ApplicationException
 from common.utils.is_reference import is_link, is_reference
-from common.utils.resolve_reference import ResolvedAddress, resolve_address
+from common.utils.resolve_address import ResolvedAddress, resolve_address
 from storage.data_source_class import DataSource
