+++ src/tests/unit/common/utils/test_reference_to_reference_items.py
@@ -1,6 +1,6 @@
 import unittest
 
-from common.utils.resolve_reference import (
+from common.utils.resolve_address import (
     AttributeItem,
     IdItem,
     QueryItem,
